requirements.txt
numpy
torch
torchvision
kagglehub
pillow
matplotlib
pandas
tqdm
opencv-python
git+https://github.com/jacobgil/pytorch-grad-cam